September 2025

+++ recursion/non_attacking_queens.cc
@@ -10,8 +10,8 @@ std::vector<std::vector<int32_t>> FindQueensPosition(int32_t n) {
 
   // Test if newly placed queens will conflict any earlier queens placed before.
   auto is_valid = [&](const std::vector<int32_t>& col_placement) {
-    const size_t row_id = col_placement.size() - 1;
-    for (size_t row = 0; row < row_id; ++row) {
+    const int32_t row_id = col_placement.size() - 1;
+    for (int32_t row = 0; row < row_id; ++row) {
       const int32_t diff = std::abs(col_placement[row] - col_placement[row_id]);
       if (diff == 0 || diff == row_id) {
         // A column or diagonal is violated
@@ -21,13 +21,13 @@ std::vector<std::vector<int32_t>> FindQueensPosition(int32_t n) {
     return true;
   };
 
-  std::function<void(size_t, std::vector<int32_t>*)> solve =
-      [&](size_t row, std::vector<int32_t>* col_placement) {
+  std::function<void(int32_t, std::vector<int32_t>*)> solve =
+      [&](int32_t row, std::vector<int32_t>* col_placement) {
         if (row == n) {
           // All queens are legally placed
           result.emplace_back(*col_placement);
         } else {
-          for (size_t col = 0; col < n; ++col) {
+          for (int32_t col = 0; col < n; ++col) {
             col_placement->emplace_back(col);
             if (is_valid(*col_placement)) {
               solve(row + 1, col_placement);
